HDDS-9710. Missing snapshot entries list Snapshot under a bucket API
